requirements.txt
Flask==3.1.0
requests==2.31.0
urllib3==2.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
telebot==0.0.5
colorama==0.4.6
tzdata==2025.1
httpx==0.27.0
